pyproject.toml
[project]
name = "proxasaurus"
version = "0.1.0"
description = "Full-stack infrastructure MCP server — Proxmox, Kubernetes, Tailscale"
requires-python = ">=3.11"
dependencies = [
    "mcp[cli]>=1.0.0",
    "requests>=2.28.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
proxasaurus = "proxasaurus.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/proxasaurus"]
